Support Timezones by accepting a Time-like object Nickel.time_class
Similar to Chronic's time_class configuration, I've added an accessor which
allows the user to set a Time-like object for Nickel to use when parsing.

This just gets around the hassle of passing in a zoned Time object to
Nickel.parse every time.

Updated the specs to reflect newest changes.